// src/app/api/model/getRecurringDepositProductsProductIdResponse.ts
export interface GetRecurringDepositProductsProductIdResponse { 
    accountingMappings?: GetRecurringDepositProductsProductIdAccountingMappings;
    activeChart?: GetRecurringDepositProductsProductIdActiveChart;
    currency?: GetRecurringDepositProductsProductIdCurrency;
    description?: string;
    feeToIncomeAccountMappings?: Set<GetRecurringDepositProductsProductIdFeeToIncomeAccountMappings>;
    id?: number;
    interestCalculationDaysInYearType?: GetRecurringDepositProductsInterestCalculationDaysInYearType;
    interestCalculationType?: GetRecurringDepositProductsInterestCalculationType;
    interestCompoundingPeriodType?: GetRecurringDepositProductsProductIdInterestCompoundingPeriodType;
    interestPostingPeriodType?: GetRecurringDepositProductsInterestPostingPeriodType;
    maxDepositTerm?: number;
    maxDepositTermType?: GetRecurringDepositProductsProductIdMaxDepositTermType;
    minDepositTerm?: number;
    minDepositTermType?: GetRecurringDepositProductsProductIdMinDepositTermType;
    name?: string;
    penaltyToIncomeAccountMappings?: Set<GetRecurringDepositProductsProductIdPenaltyToIncomeAccountMappings>;
    preClosurePenalApplicable?: boolean;
    preClosurePenalInterest?: number;
    preClosurePenalInterestOnType?: GetRecurringDepositProductsProductIdPreClosurePenalInterestOnType;
    recurringDepositFrequency?: number;
    recurringDepositFrequencyType?: GetRecurringDepositProductsRecurringDepositFrequencyType;
    shortName?: string;
}